Delete unused WebSocket configuration file and move CORSFilter, which is currently not used/configured to the configuration package
