--- samples/core/src/main/java/io/oasp/gastronomy/restaurant/general/configuration/AppWebSocketConfig.java
@@ -1,31 +1,0 @@
-package io.oasp.gastronomy.restaurant.general.configuration;
-
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.AbstractWebSocketMessageBrokerConfigurer;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-
-/**
- * ..
- *
- * @author rose
- */
-// @Configuration
-// @EnableWebSocketMessageBroker
-public class AppWebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
-  @Override
-  public void configureMessageBroker(MessageBrokerRegistry config) {
-
-    config.enableSimpleBroker("/topic");
-    config.setApplicationDestinationPrefixes("/sample");
-  }
-
-  @Override
-  public void registerStompEndpoints(StompEndpointRegistry registry) {
-
-    registry.addEndpoint("/positions").withSockJS()
-            .setStreamBytesLimit(512 * 1024)
-            .setHttpMessageCacheSize(1000)
-            .setDisconnectDelay(10 * 1000);
-  }
-
-}
